requirements.txt
streamlit
google-genai
PyYAML
sentence-transformers
faiss-cpu
